Re-included the ASIOSDK2 folder (but only using the 2 needed headers from the SDK's common folder) Shuffled some code around and renamed files to avoid confusion with ASIOSDK files. Excluded ASIOSDK from clang. Excluded macOS build files from git.
