+++ .clang-format-ignore
@@ -1,3 +1,4 @@
 # exclude third party code from clang-format checks
 ./libs
 ./windows/nsProcess
+./windows/ASIOSDK2
