+++ .gitignore
@@ -8,7 +8,7 @@ Makefile
 .vs
 .cproject
 .project
-.settings
+.settings           
 .idea
 .vscode
 .cache
@@ -43,6 +43,10 @@ distributions/fluidsynth*
 distributions/jamulus.desktop
 distributions/jamulus-server.desktop
 .xcode
+*.xcodeproj
+._*
+*.bak
+*.log
 Debug-iphoneos/
 Jamulus.xcodeproj
 jamulus_plugin_import.cpp
