+++ windows/asiodrivers.cpp
@@ -1,3 +1,12 @@
+// ******************************************************************
+// asiodrivers.cpp
+//  ASIO device selection classes by pgScorpio
+//  Obtains the list of available ASIO devices on the system and
+//  selects the current ASIO device to be used.
+//  Also implements the ASIO functions defined in asio.h to access
+//  the currently selected device.
+// ******************************************************************
+
 #include "asiodrivers.h"
 
 #include <string.h>
@@ -67,6 +76,12 @@ static bool findDrvPath ( char* clsidstr, char* dllpath, int dllpathsize )
 
     return pathfound;
 }
+
+//============================================================================
+// The current ASIO interface
+//============================================================================
+
+IASIO* pAsioDriver = NO_ASIO_DRIVER;
 
 //============================================================================
 //  CAsioDummy: a Dummy ASIO driver
@@ -693,8 +708,86 @@ bool CAsioDrivers::loadDriver ( char* name )
 }
 
 //============================================================================
-//  The global variables
+//  The global instance
 //============================================================================
 
 CAsioDrivers AsioDrivers;
-IASIO*       pAsioDriver = NO_ASIO_DRIVER;
+
+//============================================================================
+// The ASIO functions declared in asio.h
+//  These functions will access the currently selected device.
+//============================================================================
+
+ASIOError ASIOInit ( ASIODriverInfo* info )
+{
+    if ( pAsioDriver == NULL )
+    {
+        // pAsioDriver should never be NULL!
+        // Should Point to NO_ASIO_DRIVER if no driver loaded
+        pAsioDriver = NO_ASIO_DRIVER;
+    }
+
+    pAsioDriver->getDriverName ( info->name );
+
+    info->driverVersion = pAsioDriver->getDriverVersion();
+
+    if ( pAsioDriver->init ( info->sysRef ) )
+    {
+        info->errorMessage[0] = 0;
+        return ASE_OK;
+    }
+    else
+    {
+        pAsioDriver->getErrorMessage ( info->errorMessage );
+        return ( pAsioDriver == NO_ASIO_DRIVER ) ? ASE_NotPresent : ASE_HWMalfunction;
+    }
+}
+
+ASIOError ASIOExit ( void )
+{
+    AsioDrivers.closeDriver();
+    return ASE_OK;
+}
+
+ASIOError ASIOStart ( void ) { return pAsioDriver->start(); }
+
+ASIOError ASIOStop ( void ) { return pAsioDriver->stop(); }
+
+ASIOError ASIOGetChannels ( long* numInputChannels, long* numOutputChannels )
+{
+    return pAsioDriver->getChannels ( numInputChannels, numOutputChannels );
+}
+
+ASIOError ASIOGetLatencies ( long* inputLatency, long* outputLatency ) { return pAsioDriver->getLatencies ( inputLatency, outputLatency ); }
+
+ASIOError ASIOGetBufferSize ( long* minSize, long* maxSize, long* preferredSize, long* granularity )
+{
+    return pAsioDriver->getBufferSize ( minSize, maxSize, preferredSize, granularity );
+}
+
+ASIOError ASIOCanSampleRate ( ASIOSampleRate sampleRate ) { return pAsioDriver->canSampleRate ( sampleRate ); }
+
+ASIOError ASIOGetSampleRate ( ASIOSampleRate* currentRate ) { return pAsioDriver->getSampleRate ( currentRate ); }
+
+ASIOError ASIOSetSampleRate ( ASIOSampleRate sampleRate ) { return pAsioDriver->setSampleRate ( sampleRate ); }
+
+ASIOError ASIOGetClockSources ( ASIOClockSource* clocks, long* numSources ) { return pAsioDriver->getClockSources ( clocks, numSources ); }
+
+ASIOError ASIOSetClockSource ( long reference ) { return pAsioDriver->setClockSource ( reference ); }
+
+ASIOError ASIOGetSamplePosition ( ASIOSamples* sPos, ASIOTimeStamp* tStamp ) { return pAsioDriver->getSamplePosition ( sPos, tStamp ); }
+
+ASIOError ASIOGetChannelInfo ( ASIOChannelInfo* info ) { return pAsioDriver->getChannelInfo ( info ); }
+
+ASIOError ASIOCreateBuffers ( ASIOBufferInfo* bufferInfos, long numChannels, long bufferSize, ASIOCallbacks* callbacks )
+{
+    return pAsioDriver->createBuffers ( bufferInfos, numChannels, bufferSize, callbacks );
+}
+
+ASIOError ASIODisposeBuffers ( void ) { return pAsioDriver->disposeBuffers(); }
+
+ASIOError ASIOControlPanel ( void ) { return pAsioDriver->controlPanel(); }
+
+ASIOError ASIOFuture ( long selector, void* opt ) { return pAsioDriver->future ( selector, opt ); }
+
+ASIOError ASIOOutputReady ( void ) { return pAsioDriver->outputReady(); }
